sun_cjeck

+++ ss_oop/ss_oop/CorpCeresc.cpp
@@ -1,5 +1,5 @@
 #include "CorpCeresc.h"
 #include <iostream>
-CorpCeresc::CorpCeresc(std::string nume, double masa, double diametru, double x, double y)
+CorpCeresc::CorpCeresc(std::string nume, float masa, float diametru, float x, float y)
 	: nume(nume), masa(masa), diametru(diametru), x(x), y(y) {
 }
